move functions for tree optimization into forest.{c,h}pp
